+++ CHANGELOG.md
@@ -51,8 +51,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `call_listeners` to call all listeners for an event based on the passed args
   and kwargs. This will call them directly and not utilise the message-broker
   unlike `dispatch_event`
-- Added `HTTPRateLimitError` for receiving http rate-limits (429) and parameter
+- `HTTPRateLimitError` for receiving http rate-limits (429) and parameter
   retry_on_rate_limit to raw_request()
+- Parameter `remove_listeners` to `HivenClient.close()`, which will, if set to
+  True, remove all listeners created using @client.event(),
+  add_multi_listener() and add_single_listener()
 
 ### Changed
 - Rewrite of the base structure